Moved usePreviewRange tests together with other DateRangeCalendar tests
